+++ example/lib/data/user.dart
@@ -1,0 +1,48 @@
+class User {
+  List<UserInfo> userInfo = [];
+
+  void initData(int size) {
+    userInfo.clear();
+    for (int i = 0; i < size; i++) {
+      userInfo.add(UserInfo(
+          "User_$i", i % 3 == 0, '+001 9999 9999', '2019-01-01', 'N/A'));
+    }
+  }
+
+  ///
+  /// Single sort, sort Name's id
+  void sortName(bool isAscending) {
+    userInfo.sort((a, b) {
+      int aId = int.tryParse(a.name.replaceFirst('User_', '')) ?? 0;
+      int bId = int.tryParse(b.name.replaceFirst('User_', '')) ?? 0;
+      return (aId - bId) * (isAscending ? 1 : -1);
+    });
+  }
+
+  ///
+  /// sort with Status and Name as the 2nd Sort
+  void sortStatus(bool isAscending) {
+    userInfo.sort((a, b) {
+      if (a.status == b.status) {
+        int aId = int.tryParse(a.name.replaceFirst('User_', '')) ?? 0;
+        int bId = int.tryParse(b.name.replaceFirst('User_', '')) ?? 0;
+        return (aId - bId);
+      } else if (a.status) {
+        return isAscending ? 1 : -1;
+      } else {
+        return isAscending ? -1 : 1;
+      }
+    });
+  }
+}
+
+class UserInfo {
+  String name;
+  bool status;
+  String phone;
+  String registerDate;
+  String terminationDate;
+
+  UserInfo(this.name, this.status, this.phone, this.registerDate,
+      this.terminationDate);
+}
