+++ example/lib/main.dart
@@ -1,5 +1,10 @@
+import 'package:example/data/user.dart';
+import 'package:example/simple_rtl_table.dart';
+import 'package:example/simple_table.dart';
+import 'package:example/simple_table_refresh_load.dart';
+import 'package:example/simple_table_scroll_style.dart';
+import 'package:example/simple_table_sort.dart';
 import 'package:flutter/material.dart';
-import 'package:horizontal_data_table/horizontal_data_table.dart';
 
 void main() => runApp(MyApp());
 
@@ -26,23 +31,10 @@ class MyHomePage extends StatefulWidget {
 }
 
 class _MyHomePageState extends State<MyHomePage> {
-  HDTRefreshController _hdtRefreshController = HDTRefreshController();
-
-  // ignore: unused_field
-  ScrollController _verticalScrollController = ScrollController();
-  // ignore: unused_field
-  ScrollController _horizontalScrollController = ScrollController();
-
-  static const int sortName = 0;
-  static const int sortStatus = 1;
-  bool isAscending = true;
-  int sortType = sortName;
-
-  bool isLTRmode = true;
-
+  User _user = User();
   @override
   void initState() {
-    user.initData(100);
+    _user.initData(100);
     super.initState();
   }
 
@@ -52,299 +44,60 @@ class _MyHomePageState extends State<MyHomePage> {
       appBar: AppBar(
         title: Text(widget.title),
       ),
-      body: _getBodyWidget(),
+      body: Center(
+        child: Column(
+          children: [
+            _getRouteButton(
+              'Simple Table',
+              SimpleTablePage(
+                user: _user,
+              ),
+            ),
+            _getRouteButton(
+              'RTL Table',
+              SimpleTableRTLPage(
+                user: _user,
+              ),
+            ),
+            _getRouteButton(
+              'Pull-to-refresh Table',
+              SimpleTableRefreshLoadPage(
+                user: _user,
+              ),
+            ),
+            _getRouteButton(
+              'Customize Scroll Related Table',
+              SimpleTableScrollStylePage(
+                user: _user,
+              ),
+            ),
+            _getRouteButton(
+              'Sortable Table',
+              SimpleTableSortPage(
+                user: _user,
+              ),
+            ),
+          ],
+        ),
+      ),
     );
   }
 
-  Widget _getBodyWidget() {
-    return Container(
-      child: Column(
-        children: [
-          Wrap(
-            spacing: 4.0,
-            children: [
-              ElevatedButton(
-                  onPressed: () {
-                    setState(() {
-                      isLTRmode = !isLTRmode;
-                    });
-                  },
-                  child: Text(isLTRmode
-                      ? 'change to right-to-left'
-                      : 'change to left-to-right')),
-            ],
+  Widget _getRouteButton(String label, Widget page) {
+    return ElevatedButton(
+      onPressed: () {
+        Navigator.push(
+          context,
+          MaterialPageRoute(
+            builder: (context) {
+              return page;
+            },
           ),
-
-          /// horizontal data table must declaire a finite height
-          /// either fixed height or using expanded
-          Expanded(
-            child: isLTRmode ? _getLTRTable() : _getRTLTable(),
-          ),
-        ],
+        );
+      },
+      child: Text(
+        label,
       ),
     );
   }
-
-  /// left to right table
-  /// same setting as the previous version
-  Widget _getLTRTable() {
-    return HorizontalDataTable(
-      leftHandSideColumnWidth: 100,
-      rightHandSideColumnWidth: 600,
-      isFixedHeader: true,
-      headerWidgets: _getTitleWidget(),
-      leftSideItemBuilder: _generateFirstColumnRow,
-      rightSideItemBuilder: _generateRightHandSideColumnRow,
-      itemCount: user.userInfo.length,
-      rowSeparatorWidget: const Divider(
-        color: Colors.black54,
-        height: 1.0,
-        thickness: 0.0,
-      ),
-      leftHandSideColBackgroundColor: Color(0xFFFFFFFF),
-      rightHandSideColBackgroundColor: Color(0xFFFFFFFF),
-      onScrollControllerReady: (vertical, horizontal) {
-        _verticalScrollController = vertical;
-        _horizontalScrollController = horizontal;
-      },
-      verticalScrollbarStyle: const ScrollbarStyle(
-        thumbColor: Colors.yellow,
-        isAlwaysShown: true,
-        thickness: 4.0,
-        radius: Radius.circular(5.0),
-      ),
-      horizontalScrollbarStyle: const ScrollbarStyle(
-        thumbColor: Colors.red,
-        isAlwaysShown: true,
-        thickness: 4.0,
-        radius: Radius.circular(5.0),
-      ),
-      enablePullToRefresh: true,
-      fixedSidePlaceHolderRefreshIndicator: PlaceholderHeader(),
-      refreshIndicator: ClassicHeader(),
-      refreshIndicatorHeight: 60,
-      onRefresh: () async {
-        //Do sth
-        await Future.delayed(const Duration(milliseconds: 500));
-        _hdtRefreshController.refreshCompleted();
-      },
-      enablePullToLoadNewData: true,
-      fixedSidePlaceHolderLoadIndicator: PlaceholderFooter(),
-      loadIndicator: const ClassicFooter(),
-      onLoad: () async {
-        //Do sth
-        await Future.delayed(const Duration(milliseconds: 500));
-        _hdtRefreshController.loadComplete();
-      },
-      htdRefreshController: _hdtRefreshController,
-    );
-  }
-
-  /// right to left table
-  /// experimental, as i am not so familiar with the Arabic app UX,
-  /// please feel free to comment and contribute with the rtl behaviour
-  Widget _getRTLTable() {
-    return HorizontalDataTable.rtl(
-      rightHandSideColumnWidth: 100,
-      leftHandSideColumnWidth: 600,
-      isFixedHeader: true,
-      headerWidgets: _getTitleWidget(),
-      rightSideItemBuilder: _generateFirstColumnRow,
-      leftSideItemBuilder: _generateRightHandSideColumnRow,
-      itemCount: user.userInfo.length,
-      rowSeparatorWidget: const Divider(
-        color: Colors.black54,
-        height: 1.0,
-        thickness: 0.0,
-      ),
-      rightHandSideColBackgroundColor: Color(0xFFFFFFFF),
-      leftHandSideColBackgroundColor: Color(0xFFFFFFFF),
-      onScrollControllerReady: (vertical, horizontal) {
-        _verticalScrollController = vertical;
-        _horizontalScrollController = horizontal;
-      },
-      verticalScrollbarStyle: const ScrollbarStyle(
-        thumbColor: Colors.yellow,
-        isAlwaysShown: true,
-        thickness: 4.0,
-        radius: Radius.circular(5.0),
-      ),
-      horizontalScrollbarStyle: const ScrollbarStyle(
-        thumbColor: Colors.red,
-        isAlwaysShown: true,
-        thickness: 4.0,
-        radius: Radius.circular(5.0),
-      ),
-      enablePullToRefresh: true,
-      refreshIndicator: const WaterDropHeader(),
-      fixedSidePlaceHolderRefreshIndicator: PlaceholderHeader(),
-      refreshIndicatorHeight: 60,
-      onRefresh: () async {
-        //Do sth
-        await Future.delayed(const Duration(milliseconds: 500));
-        _hdtRefreshController.refreshCompleted();
-      },
-      enablePullToLoadNewData: true,
-      loadIndicator: const ClassicFooter(),
-      fixedSidePlaceHolderLoadIndicator: PlaceholderFooter(),
-      onLoad: () async {
-        //Do sth
-        await Future.delayed(const Duration(milliseconds: 500));
-        _hdtRefreshController.loadComplete();
-      },
-      htdRefreshController: _hdtRefreshController,
-    );
-  }
-
-  List<Widget> _getTitleWidget() {
-    return [
-      TextButton(
-        style: TextButton.styleFrom(
-          padding: EdgeInsets.zero,
-        ),
-        child: _getTitleItemWidget(
-            'Name' + (sortType == sortName ? (isAscending ? '↓' : '↑') : ''),
-            100),
-        onPressed: () {
-          sortType = sortName;
-          isAscending = !isAscending;
-          user.sortName(isAscending);
-          setState(() {});
-        },
-      ),
-      TextButton(
-        style: TextButton.styleFrom(
-          padding: EdgeInsets.zero,
-        ),
-        child: _getTitleItemWidget(
-            'Status' +
-                (sortType == sortStatus ? (isAscending ? '↓' : '↑') : ''),
-            100),
-        onPressed: () {
-          sortType = sortStatus;
-          isAscending = !isAscending;
-          user.sortStatus(isAscending);
-          setState(() {});
-        },
-      ),
-      _getTitleItemWidget('Phone', 200),
-      _getTitleItemWidget('Register', 100),
-      _getTitleItemWidget('Termination', 200),
-    ];
-  }
-
-  Widget _getTitleItemWidget(String label, double width) {
-    return Container(
-      child: Text(label, style: TextStyle(fontWeight: FontWeight.bold)),
-      width: width,
-      height: 56,
-      padding: EdgeInsets.fromLTRB(5, 0, 0, 0),
-      alignment: Alignment.centerLeft,
-    );
-  }
-
-  Widget _generateFirstColumnRow(BuildContext context, int index) {
-    return Container(
-      child: Text(user.userInfo[index].name),
-      width: 100,
-      height: 52,
-      padding: EdgeInsets.fromLTRB(5, 0, 0, 0),
-      alignment: Alignment.centerLeft,
-    );
-  }
-
-  Widget _generateRightHandSideColumnRow(BuildContext context, int index) {
-    return Row(
-      children: <Widget>[
-        Container(
-          child: Row(
-            children: <Widget>[
-              Icon(
-                  user.userInfo[index].status
-                      ? Icons.notifications_off
-                      : Icons.notifications_active,
-                  color:
-                      user.userInfo[index].status ? Colors.red : Colors.green),
-              Text(user.userInfo[index].status ? 'Disabled' : 'Active')
-            ],
-          ),
-          width: 100,
-          height: 52,
-          padding: EdgeInsets.fromLTRB(5, 0, 0, 0),
-          alignment: Alignment.centerLeft,
-        ),
-        Container(
-          child: Text(user.userInfo[index].phone),
-          width: 200,
-          height: 52,
-          padding: EdgeInsets.fromLTRB(5, 0, 0, 0),
-          alignment: Alignment.centerLeft,
-        ),
-        Container(
-          child: Text(user.userInfo[index].registerDate),
-          width: 100,
-          height: 52,
-          padding: EdgeInsets.fromLTRB(5, 0, 0, 0),
-          alignment: Alignment.centerLeft,
-        ),
-        Container(
-          child: Text(user.userInfo[index].terminationDate),
-          width: 200,
-          height: 52,
-          padding: EdgeInsets.fromLTRB(5, 0, 0, 0),
-          alignment: Alignment.centerLeft,
-        ),
-      ],
-    );
-  }
-}
-
-User user = User();
-
-class User {
-  List<UserInfo> userInfo = [];
-
-  void initData(int size) {
-    for (int i = 0; i < size; i++) {
-      userInfo.add(UserInfo(
-          "User_$i", i % 3 == 0, '+001 9999 9999', '2019-01-01', 'N/A'));
-    }
-  }
-
-  ///
-  /// Single sort, sort Name's id
-  void sortName(bool isAscending) {
-    userInfo.sort((a, b) {
-      int aId = int.tryParse(a.name.replaceFirst('User_', '')) ?? 0;
-      int bId = int.tryParse(b.name.replaceFirst('User_', '')) ?? 0;
-      return (aId - bId) * (isAscending ? 1 : -1);
-    });
-  }
-
-  ///
-  /// sort with Status and Name as the 2nd Sort
-  void sortStatus(bool isAscending) {
-    userInfo.sort((a, b) {
-      if (a.status == b.status) {
-        int aId = int.tryParse(a.name.replaceFirst('User_', '')) ?? 0;
-        int bId = int.tryParse(b.name.replaceFirst('User_', '')) ?? 0;
-        return (aId - bId);
-      } else if (a.status) {
-        return isAscending ? 1 : -1;
-      } else {
-        return isAscending ? -1 : 1;
-      }
-    });
-  }
-}
-
-class UserInfo {
-  String name;
-  bool status;
-  String phone;
-  String registerDate;
-  String terminationDate;
-
-  UserInfo(this.name, this.status, this.phone, this.registerDate,
-      this.terminationDate);
 }
