+++ lib/scroll/linked_scroll_controller/linked_scroll_controller.dart
@@ -206,9 +206,7 @@ class _LinkedScrollPosition extends ScrollPositionWithSingleContext {
           context: context,
           initialPixels: initialPixels,
           oldPosition: oldPosition,
-        ) {
-    assert(owner != null);
-  }
+        );
 
   final _LinkedScrollController owner;
